+++ ClassLibrary1/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace ClassLibrary1
+{
+    public class Class1
+    {
+
+    }
+}
