+++ WebApplication1/Properties/launchSettings.json
@@ -4,19 +4,19 @@
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
     "iisExpress": {
-      "applicationUrl": "http://localhost:11185",
-      "sslPort": 44315
+      "applicationUrl": "http://localhost:24099",
+      "sslPort": 44335
     }
   },
   "profiles": {
-    "api": {
+    "WebApplication1": {
       "commandName": "Project",
       "launchBrowser": true,
       "launchUrl": "swagger",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       },
-      "applicationUrl": "https://localhost:7047;http://localhost:5047",
+      "applicationUrl": "https://localhost:7107;http://localhost:5107",
       "dotnetRunMessages": true
     },
     "IIS Express": {
