--- api/api.v3.ncrunchproject
@@ -1,5 +1,0 @@
-﻿<ProjectConfiguration>
-  <Settings>
-    <HiddenComponentWarnings />
-  </Settings>
-</ProjectConfiguration>
