--- api/appsettings.Development.json
@@ -1,12 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Trace",
-      "Microsoft.AspNetCore": "Warning",
-      "Microsoft.Extensions.Hosting.Internal": "Warning",
-      "ncrunch.MockAuthenticationHandler": "Warning",
-      "ncrunch.TestAuthHandler": "Warning",
-      "ncrunch.TaskRunner": "Warning"
-    }
-  }
-}
